refactor(stores): git.ts를 패키지로 승격 + 타입 추출
1,448줄 → types.ts 499줄(타입 34종 + 타임아웃 상수 3개 + appendWorktreePath
헬퍼) + index.ts 1,013줄. index.ts는 아직 800줄 한도를 초과한다 — 액션 63개를
도메인 모듈로 승격하는 작업은 Task 5에서 처리한다.

interface GitState에 export 추가(패키지 내부 export, Task 5 도메인 모듈이
참조). 소비자 재노출은 브리프의 22개 + tsc가 잡아낸 누락 4개
(PruneCandidate·PruneSkipReason·PruneSkipped·ResolutionStrategy) = 26개
타입 + useGitStore 훅.

표면 스냅샷(git.surface.test.ts) + 기존 테스트(git.test.ts, 60개) 통과.
tsc/eslint/vitest(200 files, 4354 tests)/build 전체 통과.
